Copulas in wh-questions
"How old is Brian Eno": "How old" heads "is Brian Eno"
"How white did she paint the wall": "did she paint the wall" heads "How white"

The category of "How" is the same in both cases, so how can we tell if its
second argument (s:q/(s:adj\np)) is inverting or not?

We can tell by the annotation of the s:adj\np part: if it is annotated "flip",
s:q/(s:adj\np) is a copula category, and we should annotate is as noninverting
so it ends up depending on "How old".

Problem: we don't know if the s:adj\np part is already annotated.

Solution: use Prolog co-routining to wait until it is.

Argh, this is getting slightly messy. But so far it works!