Towards a less hacky treatment of modifiers
Our old method of just unifying result and argument of modifier categories is
not compatible with the new approach of ignoring unconsumed arguments of
arguments. So for now, we do not copy annotations from arguments to modifiers.
We will have to do this somehow to ensure correct dependencies. But for now, at
least the infamous "Ms Mohn" example is converted to a dependency tree.