handle_roles/3
While handle_roles/2 just throws roles away and can thus only be applied in
situations where we have given up on handling any potential roles,
handle_roles/3 aims to be more broadly applicable: if there is a role, we
assign it to the argument and strip it off the rolelist, if not, we quietly
succeed without any new bindings. We still have to take care to not call it
when an argument must be skipped for role assignment (this is notably the case
for PP arguments), but other than that, we should be able to use it in every
rule without thinking.