A new way to map event-token pairs to roletags
Sometimes, the same participant fills multiple roles for the same event, e.g.,

1. in sentences like "I haven't been to Boston since 2013", where "2013" fills
   the Time and Start roles for "been".
2. in reflexive and reciprocal sentences like "Tom and Mary detest each other",
   where "Tom and Mary" and "each other" corefer and fill both the Experiencer
   and the Stimulus role for "detest".

The old heuristic was to always pick the *last* encountered role.

The new heuristic is to optimize for the minimum distance between the token
introducing the participant and the token introducing the role.

The local `roletag` method has become more complex to implement this. On the
other hand, we were able to get rid of the nested dictionary `pas` and replace
it with a list `roless` that is simpler and more similar to `refss`.